Create kernel simulation project

+++ include/kernel/core.h
@@ -158,9 +158,9 @@ _Noreturn void cpu_halt();
 
 /* -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-= */
 
-bio_t *bio_create(inode_t *ino, int flags, int block, int offset);
-void *bio_access(bio_t *io, int lba);
-void bio_clean(bio_t *io, int lba);
+bio_t *bio_create(inode_t *ino, int flags, int block, size_t offset);
+void *bio_access(bio_t *io, size_t lba);
+void bio_clean(bio_t *io, size_t lba);
 void bio_destroy(bio_t *io);
 
 /* -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-= */
